Some change to reduce cpp usage in the Main file.
- Made separate modules for each ghc version in the
  hopes that maintaining them would become easier
  than the many nested if, else we have in cpp right now.
